+++ src/main/java/com/refinedmods/refinedstorage/screen/widget/ScrollbarWidget.java
@@ -72,7 +72,7 @@ public class ScrollbarWidget implements IGuiEventListener {
 
         if (button == 0 && RenderUtils.inBounds(x, y, width, height, mx, my)) {
             // Prevent accidental scrollbar click after clicking recipe transfer button
-            if (JeiIntegration.isLoaded() && System.currentTimeMillis() - GridRecipeTransferHandler.LAST_TRANSFER_TIME <= GridRecipeTransferHandler.TRANSFER_SCROLLBAR_DELAY_MS) {
+            if (JeiIntegration.isLoaded() && System.currentTimeMillis() - GridRecipeTransferHandler.lastTransferTime <= GridRecipeTransferHandler.TRANSFER_SCROLLBAR_DELAY_MS) {
                 return false;
             }